Vertical Order traversal of the tree
The key idea in this traversal is the constraint. You can take that the constraint is 20 nodes i.e 20 nodes can be on the left side and 20 node on the right side. Thus it is okay to be declare the arraylist of the size 40 so that it can cover both the sides. 
The main point is that the current arraylist candidate would be x then it should be passed with the pointer x+1 and x-1 to cover the right and the left subtree. That's all.